restructure bulk edit service, https://github.com/microsoft/vscode/issues/91383
create separate files for text and file operation
